pyproject.toml
[project]
name = "bizniz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.16.0",
    "fastapi>=0.118.0",
    "google-cloud-vision>=3.10.2",
    "google-generativeai>=0.8.5",
    "mutagen>=1.47.0",
    "pocketbase>=0.15.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "stripe>=13.0.1",
]
